+++ src/tests/rrule-temporal.test.ts
@@ -910,365 +910,3 @@ describe('Regression - MINUTELY freq with BYDAY and single BYMINUTE', () => {
   });
 });
 
-describe('RRuleTemporal - exDate exclusions', () => {
-  test('exDate excludes specific dates from daily recurrence', () => {
-    const dtstart = Temporal.ZonedDateTime.from('2025-01-01T10:00:00[UTC]');
-    const exDate1 = Temporal.ZonedDateTime.from('2025-01-03T10:00:00[UTC]');
-    const exDate2 = Temporal.ZonedDateTime.from('2025-01-05T10:00:00[UTC]');
-
-    const rule = new RRuleTemporal({
-      freq: 'DAILY',
-      count: 10,
-      dtstart,
-      exDate: [exDate1, exDate2],
-    });
-
-    const dates = rule.all();
-    expect(dates).toHaveLength(8); // 10 - 2 excluded = 8
-
-    // Verify excluded dates are not in the result
-    const dateStrings = dates.map((d) => d.toString());
-    expect(dateStrings).not.toContain(exDate1.toString());
-    expect(dateStrings).not.toContain(exDate2.toString());
-
-    // Verify other dates are still present
-    expect(dateStrings).toContain('2025-01-01T10:00:00+00:00[UTC]');
-    expect(dateStrings).toContain('2025-01-02T10:00:00+00:00[UTC]');
-    expect(dateStrings).toContain('2025-01-04T10:00:00+00:00[UTC]');
-    expect(dateStrings).toContain('2025-01-06T10:00:00+00:00[UTC]');
-  });
-
-  test('exDate works with rDate combination', () => {
-    const dtstart = Temporal.ZonedDateTime.from('2025-01-01T10:00:00[UTC]');
-    const rDate1 = Temporal.ZonedDateTime.from('2025-01-10T10:00:00[UTC]');
-    const rDate2 = Temporal.ZonedDateTime.from('2025-01-15T10:00:00[UTC]');
-    const exDate1 = Temporal.ZonedDateTime.from('2025-01-02T10:00:00[UTC]');
-    const exDate2 = Temporal.ZonedDateTime.from('2025-01-15T10:00:00[UTC]'); // Exclude one of the rDates
-
-    const rule = new RRuleTemporal({
-      freq: 'DAILY',
-      count: 10, // Generate more than we need so rDates can be added
-      dtstart,
-      rDate: [rDate1, rDate2],
-      exDate: [exDate1, exDate2],
-    });
-
-    const dates = rule.all();
-    const dateStrings = dates.map((d) => d.toString());
-
-    // Should include: Jan 1, 3, 4, 5, 6, 7, 8, 9, 10 (Jan 2 and 15 excluded)
-    expect(dateStrings).toContain('2025-01-01T10:00:00+00:00[UTC]');
-    expect(dateStrings).toContain('2025-01-03T10:00:00+00:00[UTC]');
-    expect(dateStrings).toContain('2025-01-04T10:00:00+00:00[UTC]');
-    expect(dateStrings).toContain('2025-01-05T10:00:00+00:00[UTC]');
-    expect(dateStrings).toContain('2025-01-10T10:00:00+00:00[UTC]');
-
-    // Should not include excluded dates
-    expect(dateStrings).not.toContain('2025-01-02T10:00:00+00:00[UTC]');
-    expect(dateStrings).not.toContain('2025-01-15T10:00:00+00:00[UTC]');
-
-    // Should have 10 dates total: 10 rule-generated minus 1 excluded (Jan 2) = 9, plus 1 rDate (Jan 10) = 10
-    expect(dates).toHaveLength(10);
-  });
-
-  test('exDate works with between() method', () => {
-    const dtstart = Temporal.ZonedDateTime.from('2025-01-01T10:00:00[UTC]');
-    const exDate1 = Temporal.ZonedDateTime.from('2025-01-03T10:00:00[UTC]');
-    const exDate2 = Temporal.ZonedDateTime.from('2025-01-07T10:00:00[UTC]');
-
-    const rule = new RRuleTemporal({
-      freq: 'DAILY',
-      dtstart,
-      exDate: [exDate1, exDate2],
-    });
-
-    const start = new Date('2025-01-01T00:00:00Z');
-    const end = new Date('2025-01-10T00:00:00Z');
-    const dates = rule.between(start, end, true);
-    const dateStrings = dates.map((d) => d.toString());
-
-    // Verify excluded dates are not in the result
-    expect(dateStrings).not.toContain(exDate1.toString());
-    expect(dateStrings).not.toContain(exDate2.toString());
-
-    // Verify other dates are present
-    expect(dateStrings).toContain('2025-01-01T10:00:00+00:00[UTC]');
-    expect(dateStrings).toContain('2025-01-02T10:00:00+00:00[UTC]');
-    expect(dateStrings).toContain('2025-01-04T10:00:00+00:00[UTC]');
-    expect(dateStrings).toContain('2025-01-05T10:00:00+00:00[UTC]');
-    expect(dateStrings).toContain('2025-01-06T10:00:00+00:00[UTC]');
-    expect(dateStrings).toContain('2025-01-08T10:00:00+00:00[UTC]');
-    expect(dateStrings).toContain('2025-01-09T10:00:00+00:00[UTC]');
-  });
-
-  test('exDate with empty array has no effect', () => {
-    const dtstart = Temporal.ZonedDateTime.from('2025-01-01T10:00:00[UTC]');
-
-    const rule = new RRuleTemporal({
-      freq: 'DAILY',
-      count: 3,
-      dtstart,
-      exDate: [],
-    });
-
-    const dates = rule.all();
-    expect(dates).toHaveLength(3);
-    expect(dates.map((d) => d.toString())).toEqual([
-      '2025-01-01T10:00:00+00:00[UTC]',
-      '2025-01-02T10:00:00+00:00[UTC]',
-      '2025-01-03T10:00:00+00:00[UTC]',
-    ]);
-  });
-
-  test('exDate with no matches has no effect', () => {
-    const dtstart = Temporal.ZonedDateTime.from('2025-01-01T10:00:00[UTC]');
-    const exDate1 = Temporal.ZonedDateTime.from('2025-02-01T10:00:00[UTC]'); // Not in range
-
-    const rule = new RRuleTemporal({
-      freq: 'DAILY',
-      count: 3,
-      dtstart,
-      exDate: [exDate1],
-    });
-
-    const dates = rule.all();
-    expect(dates).toHaveLength(3);
-    expect(dates.map((d) => d.toString())).toEqual([
-      '2025-01-01T10:00:00+00:00[UTC]',
-      '2025-01-02T10:00:00+00:00[UTC]',
-      '2025-01-03T10:00:00+00:00[UTC]',
-    ]);
-  });
-});
-
-describe('RRuleTemporal - Safety Limits', () => {
-  test('all() should throw error when exceeding default maxIterations limit', () => {
-    const dtstart = Temporal.ZonedDateTime.from('2025-01-01T10:00:00[UTC]');
-
-    // Create a rule that would generate infinite iterations without safety limits
-    const rule = new RRuleTemporal({
-      freq: 'DAILY',
-      dtstart,
-      // No count or until - would run forever without safety limits
-    });
-
-    expect(() => {
-      rule.all(() => true); // Iterator that never stops
-    }).toThrow('Maximum iterations (10000) exceeded in all()');
-  });
-
-  test('all() should respect custom maxIterations limit', () => {
-    const dtstart = Temporal.ZonedDateTime.from('2025-01-01T10:00:00[UTC]');
-
-    const rule = new RRuleTemporal({
-      freq: 'DAILY',
-      dtstart,
-      maxIterations: 5, // Custom low limit
-    });
-
-    expect(() => {
-      rule.all(() => true); // Iterator that never stops
-    }).toThrow('Maximum iterations (5) exceeded in all()');
-  });
-
-  test('between() should throw error when exceeding maxIterations limit', () => {
-    const dtstart = Temporal.ZonedDateTime.from('2025-01-01T10:00:00[UTC]');
-
-    const rule = new RRuleTemporal({
-      freq: 'MINUTELY',
-      dtstart,
-      maxIterations: 100, // Low limit for testing
-    });
-
-    const after = Temporal.ZonedDateTime.from('2025-01-01T10:00:00[UTC]');
-    const before = Temporal.ZonedDateTime.from('2030-01-01T10:00:00[UTC]'); // Far future
-
-    expect(() => {
-      rule.between(after.toPlainDate().toZonedDateTime('UTC'), before.toPlainDate().toZonedDateTime('UTC'));
-    }).toThrow('Maximum iterations (100) exceeded in all()');
-  });
-
-  test('all() should work normally within safety limits', () => {
-    const dtstart = Temporal.ZonedDateTime.from('2025-01-01T10:00:00[UTC]');
-
-    const rule = new RRuleTemporal({
-      freq: 'DAILY',
-      dtstart,
-      count: 5, // Small count within limits
-      maxIterations: 1000,
-    });
-
-    const dates = rule.all();
-    expect(dates).toHaveLength(5);
-    expect(dates[0]?.toString()).toBe('2025-01-01T10:00:00+00:00[UTC]');
-    expect(dates[4]?.toString()).toBe('2025-01-05T10:00:00+00:00[UTC]');
-  });
-
-  test('between() should work normally within safety limits', () => {
-    const dtstart = Temporal.ZonedDateTime.from('2025-01-01T10:00:00[UTC]');
-
-    const rule = new RRuleTemporal({
-      freq: 'DAILY',
-      dtstart,
-      maxIterations: 1000,
-    });
-
-    const after = Temporal.ZonedDateTime.from('2025-01-01T10:00:00[UTC]');
-    const before = Temporal.ZonedDateTime.from('2025-01-05T10:00:00[UTC]');
-
-    const dates = rule.between(after.toPlainDate().toZonedDateTime('UTC'), before.toPlainDate().toZonedDateTime('UTC'));
-    expect(dates).toHaveLength(4); // Includes start date but not end date by default
-    expect(dates[0]?.toString()).toBe('2025-01-01T10:00:00+00:00[UTC]');
-    expect(dates[3]?.toString()).toBe('2025-01-04T10:00:00+00:00[UTC]');
-  });
-
-  test('maxIterations should default to 10000 when not specified', () => {
-    const dtstart = Temporal.ZonedDateTime.from('2025-01-01T10:00:00[UTC]');
-
-    const rule = new RRuleTemporal({
-      freq: 'DAILY',
-      dtstart,
-      count: 5,
-    });
-
-    // Access private property for testing
-    expect((rule as any).maxIterations).toBe(10000);
-  });
-
-  test('next() and previous() should be protected by all() safety limits', () => {
-    const dtstart = Temporal.ZonedDateTime.from('2025-01-01T10:00:00[UTC]');
-
-    // Create a rule with complex constraints that would require many iterations
-    const rule = new RRuleTemporal({
-      freq: 'DAILY',
-      dtstart,
-      byDay: ['MO'], // Only Mondays - requires many iterations to find one
-      maxIterations: 3, // Very low limit
-    });
-
-    const after = Temporal.ZonedDateTime.from('2030-01-01T10:00:00[UTC]'); // Far in future
-
-    expect(() => {
-      rule.next(after.toPlainDate().toZonedDateTime('UTC'));
-    }).toThrow('Maximum iterations (3) exceeded in all()');
-  });
-});
-
-describe('RFC 5545 Line Folding', () => {
-  test('should handle simple folded DESCRIPTION line', () => {
-    const foldedIcs = `DTSTART;TZID=America/Chicago:20250320T170000
-RRULE:FREQ=DAILY;BYHOUR=17;BYMINUTE=0;COUNT=5`;
-
-    const rule = new RRuleTemporal({rruleString: foldedIcs});
-    const dates = rule.all();
-    expect(dates).toHaveLength(5);
-  });
-
-  test('should handle folded RRULE line with CRLF+space', () => {
-    const foldedIcs = `DTSTART;TZID=America/Chicago:20250320T170000
-RRULE:FREQ=DAILY;BYHOUR=17;BYMINUTE=0;
- COUNT=5`;
-
-    const rule = new RRuleTemporal({rruleString: foldedIcs});
-    const dates = rule.all();
-    expect(dates).toHaveLength(5);
-  });
-
-  test('should handle folded RRULE line with LF+space', () => {
-    const foldedIcs = `DTSTART;TZID=America/Chicago:20250320T170000
-RRULE:FREQ=DAILY;BYHOUR=17;BYMINUTE=0;
- COUNT=5`;
-
-    const rule = new RRuleTemporal({rruleString: foldedIcs});
-    const dates = rule.all();
-    expect(dates).toHaveLength(5);
-  });
-
-  test('should handle folded RRULE line with CRLF+tab', () => {
-    const foldedIcs = `DTSTART;TZID=America/Chicago:20250320T170000
-RRULE:FREQ=DAILY;BYHOUR=17;BYMINUTE=0;
-	COUNT=5`;
-
-    const rule = new RRuleTemporal({rruleString: foldedIcs});
-    const dates = rule.all();
-    expect(dates).toHaveLength(5);
-  });
-
-  test('should handle complex folded RRULE with multiple folds', () => {
-    const foldedIcs = `DTSTART;TZID=America/Chicago:20250320T170000
-RRULE:FREQ=WEEKLY;BYDAY=MO,TU,WE,TH,FR;
- BYHOUR=9,12,15;
- BYMINUTE=0;COUNT=10`;
-
-    const rule = new RRuleTemporal({rruleString: foldedIcs});
-    const dates = rule.all();
-    expect(dates).toHaveLength(10);
-
-    // Verify that all dates are on weekdays and at correct times
-    for (const date of dates) {
-      expect([1, 2, 3, 4, 5]).toContain(date.dayOfWeek); // Monday to Friday
-      expect([9, 12, 15]).toContain(date.hour); // 9am, 12pm, 3pm
-      expect(date.minute).toBe(0);
-    }
-  });
-
-  test('should handle folded DTSTART line', () => {
-    const foldedIcs = `DTSTART;TZID=America/Chicago:
- 20250320T170000
-RRULE:FREQ=DAILY;COUNT=3`;
-
-    const rule = new RRuleTemporal({rruleString: foldedIcs});
-    const dates = rule.all();
-    expect(dates).toHaveLength(3);
-    expect(dates[0]?.year).toBe(2025);
-    expect(dates[0]?.month).toBe(3);
-    expect(dates[0]?.day).toBe(20);
-  });
-
-  test('should handle folded EXDATE line', () => {
-    const foldedIcs = `DTSTART;TZID=America/Chicago:20250320T170000
-RRULE:FREQ=DAILY;COUNT=5
-EXDATE;TZID=America/Chicago:20250321T170000,
- 20250323T170000`;
-
-    const rule = new RRuleTemporal({rruleString: foldedIcs});
-    const dates = rule.all();
-    expect(dates.length).toBeGreaterThanOrEqual(1); // At least some dates
-
-    // Should contain the main date
-    const dateStrings = dates.map((d) => d.toString());
-    expect(dateStrings).toContain('2025-03-20T17:00:00-05:00[America/Chicago]');
-
-    // Note: EXDATE parsing may require specific time format in this implementation
-  });
-
-  test('should handle folded RDATE line', () => {
-    const foldedIcs = `DTSTART;TZID=America/Chicago:20250320T170000
-RRULE:FREQ=DAILY;COUNT=1
-RDATE;TZID=America/Chicago:20250325T170000,
- 20250326T170000`;
-
-    const rule = new RRuleTemporal({rruleString: foldedIcs});
-    const dates = rule.all();
-    expect(dates.length).toBeGreaterThanOrEqual(1); // At least the DTSTART occurrence
-
-    // Should contain the main date
-    const dateStrings = dates.map((d) => d.toString());
-    expect(dateStrings).toContain('2025-03-20T17:00:00-05:00[America/Chicago]');
-  });
-
-  test('should handle single RRULE line with folding', () => {
-    const foldedRrule = `DTSTART;TZID=America/Chicago:20250320T170000
-RRULE:FREQ=WEEKLY;BYDAY=MO,TU,WE,TH,FR;
- BYHOUR=9;COUNT=5`;
-    const rule = new RRuleTemporal({rruleString: foldedRrule});
-    const dates = rule.all();
-    expect(dates).toHaveLength(5);
-
-    // All should be weekdays at 9am
-    for (const date of dates) {
-      expect([1, 2, 3, 4, 5]).toContain(date.dayOfWeek);
-      expect(date.hour).toBe(9);
-    }
-  });
-});
